requirements.txt
python-telegram-bot[job-queue]~=20.0
pytz~=2024.1
pydantic-settings~=2.13
groq~=0.13
openai~=1.0
SQLAlchemy~=2.0
psycopg2-binary~=2.9
supabase~=2.0
pytest~=8.0
pytest-asyncio~=0.25
